Moved AuraLogger to command package
